requirements.txt
pytesseract
gradio
opencv-python
pandas
numpy
tempfile
google.generativeai
